Category-guessing is made while categorizing transactions
When choosing a category for a new transaction(new as in from the list of transactions being added) that category might be suggested for other new transactions . This is done by putting a button below the category-chooser in the grid. The reason why the categories of these transactions aren't changed directly is that the user might have already seen and passively accepted the suggested category, so changing the category to one that is not correct might go unseen.