Mostly fixed the chosen js file
The category creation method through chosen is defective, because it only shows the create-category button in place of "no search result". So for instance if there is a category named "foobar", then "foo" can't be created because foobar would be a search-hit and prevent category-creation. Adding a native functionality in chosen to create new entries, and it's mostly been fixed in the chosen jn chosen.jquery.js now. Still have to handle it in crackshell.js