Instant automatic categorizing instead of suggestion for unviewed rows
The reason why category-suggestion-buttons are generated rather than just changing the category directly is that the user might already have viewed the row and agreed on the automatically set category, and then it is changed to something else that is incorrect. But now the categories do get set directly instead of getting suggestion-buttons, but only for rows that havent been viewed by the user, eg they have never been in view.